refactor: RoomUser hasUnread로 변경 및 에러처리 추가

+++ src/main/java/com/yuguanzhang/lumi/chat/service/ChatServiceImpl.java
@@ -9,9 +9,10 @@ import com.yuguanzhang.lumi.chat.entity.RoomUser;
 import com.yuguanzhang.lumi.chat.repository.ChatRepository;
 import com.yuguanzhang.lumi.chat.repository.RoomRepository;
 import com.yuguanzhang.lumi.chat.repository.RoomUserRepository;
+import com.yuguanzhang.lumi.common.exception.GlobalException;
+import com.yuguanzhang.lumi.common.exception.message.ExceptionMessage;
 import com.yuguanzhang.lumi.user.entity.User;
 import com.yuguanzhang.lumi.user.repository.UserRepository;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -19,7 +20,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.UUID;
 
 @Service
@@ -46,8 +46,9 @@ public class ChatServiceImpl implements ChatService {
             // 채팅방의 상대
             RoomUser senderRoomUser =
                     roomUserRepository.findByRoomUserId_RoomIdAndRoomUserId_UserIdNot(
-                                    room.getRoomId(), userId)
-                            .orElseThrow(() -> new NoSuchElementException("상대방이 존재하지 않습니다."));
+                                              room.getRoomId(), userId)
+                                      .orElseThrow(() -> new GlobalException(
+                                              ExceptionMessage.ROOM_USER_NOT_FOUND));
 
             // 상대방 정보
             User sender = senderRoomUser.getUser();
@@ -66,20 +67,24 @@ public class ChatServiceImpl implements ChatService {
         List<Chat> chatList = chatRepository.findByRoom_RoomId(roomId);
 
         chatList.forEach(chat -> {
-            if (Boolean.FALSE.equals(chat.getIsRead()) && !chat.getUser().getUserId()
-                    .equals(userId)) {
+            if (Boolean.FALSE.equals(chat.getIsRead()) && !chat.getUser()
+                                                               .getUserId()
+                                                               .equals(userId)) {
                 chat.updateIsRead();
             }
         });
 
-        // RoomUser unreadCount 초기화
+        // RoomUser hasUnread 초기화
         RoomUser roomUser =
                 roomUserRepository.findByRoomUserId_RoomIdAndRoomUserId_UserId(roomId, userId)
-                        .orElseThrow(() -> new NoSuchElementException("해당 채팅방에 사용자를 찾을 수 없습니다."));
-        roomUser.resetUnreadCount();
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.ROOM_NOT_FOUND));
+        roomUser.resetUnread();
         roomUserRepository.save(roomUser);
 
-        return chatList.stream().map(ChatsResponseDto::fromEntity).toList();
+        return chatList.stream()
+                       .map(ChatsResponseDto::fromEntity)
+                       .toList();
     }
 
     @Override
@@ -87,23 +92,60 @@ public class ChatServiceImpl implements ChatService {
     public void deleteChat(UUID userId, Long roomId, Long chatId) {
         // 채팅 메세지 조회 (에러처리 필요)
         Chat chat = chatRepository.findByRoom_RoomIdAndChatId(roomId, chatId)
-                .orElseThrow(() -> new EntityNotFoundException("해당 채팅 메세지를 찾을 수 없습니다."));
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.CHAT_NOT_FOUND));
 
-        // 커스텀 에러 처리 필요
-
-
-        // 삭제했을 때의 마지막 메세지 및 안 읽은 메세지 개수 업데이트 필요
+        if (!chat.getUser()
+                 .getUserId()
+                 .equals(userId)) {
+            throw new GlobalException(ExceptionMessage.UNAUTHORIZED_CHAT_DELETE);
+        }
 
         chatRepository.delete(chat);
+
+        // 삭제했을 때의 마지막 메세지 및 읽음 여부 업데이트 필요
+        Chat lastChat = chatRepository.findTopByRoom_RoomIdOrderByCreatedAtDesc(roomId)
+                                      .orElse(null);
+
+        List<RoomUser> roomUsers = roomUserRepository.findByRoomUserId_RoomId(roomId);
+
+        for (RoomUser ru : roomUsers) {
+            if (lastChat != null) {
+                // 마지막 메세지 갱싱
+                ru.updateLastMessage(lastChat.getContent(), lastChat.getCreatedAt());
+
+                boolean hasUnread = !lastChat.getUser()
+                                             .getUserId()
+                                             .equals(ru.getUser()
+                                                       .getUserId()) && Boolean.FALSE.equals(
+                        lastChat.getIsRead());
+
+
+                if (hasUnread) {
+                    ru.makeUnread();
+                } else {
+                    ru.resetUnread();
+                }
+            } else {
+                // 채팅방에 메세지가 더 이상 없는 경우
+                ru.updateLastMessage(null, null);
+                ru.resetUnread();
+            }
+
+            roomUserRepository.save(ru);
+        }
     }
 
     @Override
     @Transactional
     public RoomUser postChat(ChatRequestDto chatRequestDto, UUID userId, Long roomId) {
         User user = userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("사용자 정보를 찾을 수 없습니다."));
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.USER_NOT_FOUND));
+
         Room room = roomRepository.findById(roomId)
-                .orElseThrow(() -> new IllegalArgumentException("채팅방 정보를 찾을 수 없습니다."));
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.ROOM_NOT_FOUND));
 
         Chat chat = chatRequestDto.toEntity(room, user);
 
@@ -111,19 +153,23 @@ public class ChatServiceImpl implements ChatService {
 
         log.info("chat: {}", chat);
 
-        // 받는 사람 RoomUser의 lastMessage와 안읽은 메세지 수 업데이트
+        // 받는 사람 RoomUser의 lastMessage와 안읽은 메세지, 읽음 여부 업데이트
         RoomUser receiverRoomUser =
                 roomUserRepository.findByRoomUserId_RoomIdAndRoomUserId_UserIdNot(roomId, userId)
-                        .orElseThrow(() -> new NoSuchElementException("상대방이 존재하지 않습니다."));
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.ROOM_USER_NOT_FOUND));
+
 
         receiverRoomUser.updateLastMessage(chat.getContent(), chat.getCreatedAt());
-        receiverRoomUser.increaseUnreadCount();
+        receiverRoomUser.makeUnread();
 
-        // 보낸 사람 RoomUser의 lastMessage 업데이트
+        // 보낸 사람 RoomUser의 lastMessage, 읽음 여부 업데이트
         RoomUser senderRoomUser =
                 roomUserRepository.findByRoomUserId_RoomIdAndRoomUserId_UserId(roomId, userId)
-                        .orElseThrow();
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.ROOM_USER_NOT_FOUND));
         senderRoomUser.updateLastMessage(chat.getContent(), chat.getCreatedAt());
+        senderRoomUser.resetUnread();
 
         return receiverRoomUser;
     }
